restructure pocketcleaner
this adds a PocketClient type to hold API authentication data as well as the
URL to use. This makes it easier to deal with the auth data as well as mocking
things for unit testing. The external API is also clarified by making
non-essential (helper) methods un-exported and only export the one that is
actually supposed to be used.